WIP: reload configuration on CTRL+SHIFT+F1
This needs a lot of work.

The idea is to move the generation of the conf struct to separate function
read_config(). This is used by setup() to configure the initial objects.

Next CTRL+SHIFT+F1 calls reload_config() which also calls read_config()
but then applies some things.

Signal handling can't be done safely as old signals needs to be freed.
Possibly other things.